разделил Marks на htmlTags и MarkdownTags
