+++ cs/Markdown/Markdown.csproj
@@ -56,12 +56,13 @@
         <Reference Include="System.Xml" />
     </ItemGroup>
     <ItemGroup>
+        <Compile Include="HtmlTags.cs" />
+        <Compile Include="MarkdownTags.cs" />
         <Compile Include="Md.cs" />
         <Compile Include="Properties\AssemblyInfo.cs" />
         <Compile Include="RenderOperation.cs" />
         <Compile Include="Tag.cs" />
         <Compile Include="TagBuilder.cs" />
-        <Compile Include="Marks.cs" />
         <Compile Include="TagType.cs" />
     </ItemGroup>
     <ItemGroup>
